move sidePanel in oranisms
