+++ src/components/organisms/sidePanel/SidePanelController.tsx
@@ -1,10 +1,10 @@
 import React, { FunctionComponent } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { DAVIT_VERISON } from "../../DavitConstants";
-import logo from "../../icons/logo200.png";
-import { EditActions, editSelectors, Mode } from "../../slices/EditSlice";
-import { DavitSidePanelButton } from "../atomic/buttons/DavitSidePanelButton";
-import { DavitIcons } from "../common/IconSet";
+import { DAVIT_VERISON } from "../../../DavitConstants";
+import logo from "../../../icons/logo200.png";
+import { EditActions, editSelectors, Mode } from "../../../slices/EditSlice";
+import { DavitSidePanelButton } from "../../atomic/buttons/DavitSidePanelButton";
+import { DavitIcons } from "../../common/IconSet";
 
 export interface SidePanelProps {
 }
