+++ src/pages/Davit.tsx
@@ -6,7 +6,7 @@ import { useDispatch } from "react-redux";
 import { HashRouter as BrowserRouter, Route, Switch } from "react-router-dom";
 import { ErrorNotification } from "../components/common/fragments/ErrorNotification";
 import { ControlPanelController } from "../components/organisms/controllPanel/presentation/ControlPanelController";
-import { SidePanelController } from "../components/sidePanel/SidePanelController";
+import { SidePanelController } from "../components/organisms/sidePanel/SidePanelController";
 import "../css/Davit.css";
 import { ActorModelController } from "../domains/actor/ActorModelController";
 import { DataModelController } from "../domains/datamodel/DataModelController";
